# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my Laravel application",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "description": "Display a listing of the orders.",
                "operationId": "329b652e559ef498cf6599a9642b889b",
                "responses": {
                    "200": {
                        "description": "A list of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Store a new order.",
                "operationId": "505b195c00654c8a83a9f74c34e221be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "description": "Show a specific order by id.",
                "operationId": "b6735a2c34205b42126051d6505ada6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "description": "Update an order by id.",
                "operationId": "b134a8cf2d3c204ec52078f73508807b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "description": "Remove an order by id.",
                "operationId": "a23a9e05404c30afb13786d706d20562",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted"
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "description": "Order model",
                "required": [
                    "people_qty",
                    "order_date",
                    "flight_num",
                    "status",
                    "phone",
                    "email",
                    "total",
                    "arrival_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the order",
                        "type": "integer",
                        "example": 1
                    },
                    "people_qty": {
                        "description": "Number of people in the order",
                        "type": "integer",
                        "example": 3
                    },
                    "order_date": {
                        "description": "Order date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-10T10:30:00Z"
                    },
                    "flight_num": {
                        "description": "Flight number associated with the order",
                        "type": "string",
                        "example": "AB1234"
                    },
                    "status": {
                        "description": "Order status",
                        "type": "string",
                        "example": "completed"
                    },
                    "phone": {
                        "description": "Phone number of the person who placed the order",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email": {
                        "description": "Email address of the person who placed the order",
                        "type": "string",
                        "format": "email",
                        "example": "email@example.com"
                    },
                    "luggage_qty": {
                        "description": "Quantity of luggage for the order",
                        "type": "integer",
                        "example": 2
                    },
                    "total": {
                        "description": "Total price of the order",
                        "type": "number",
                        "format": "float",
                        "example": 120.5
                    },
                    "arrival_date": {
                        "description": "Arrival date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T12:00:00Z"
                    },
                    "comment": {
                        "description": "Additional comments for the order",
                        "type": "string",
                        "example": "Please arrange for a VIP transfer."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}